AL (copy-bfs): using arrays instead of struct.

+++ gen.h
@@ -10,7 +10,7 @@
 
 
 typedef int32_t vid_t;
-typedef int32_t eid_t;
+typedef int64_t eid_t;
 typedef double weight_t;
 
 typedef std::pair<vid_t, vid_t> pvv;
